requirements.txt
pandas
numpy
scipy
scikit-learn
matplotlib
tensorflow
aiohttp
statsmodels
PyYAML
seaborn
oandapyV20
pyTelegramBotAPI
python-dateutil
histdata